specify sprockets 2.11.0
As of sprockets 2.12.0 there seems to be some incompatibility with *.scss files. Possibly an incompatibility with bootstrap-sass, or sass-rails, or maybe just the asset pipeline in general. It produces a stack dump instead of rendering the page when executing the stylesheet_link_tag function. 

Locking sprockets down to 2.11.0 seems to fix it.